Friendly fallback if artefact.format is not known
If we don't understand what to do with the artefact format handed to us from the api, just render a plain base document without any metadata. In order to handle firstly allow the presenter to deal with a nil schema. Also lets provide some sensible defaults when the api response doesn't contain any headers so the view doesn't need to check for them before rendering.