Link back to pre filtered finder
When viewing a document, it's useful for the metadata values to link
back to the Finder with that filter pre selected.

This commit relies on the corresponding change in gds-api-adapters which
alters the metadata hash to have a slug as the key and inside that
contain the label and an array of values, each of those containing a
label and a slug.