Add optional beta banner
As some agencies haven't fully transitioned yet we want to show that the
main source for the documents is the agency website.

This commit adds the beta banner to be displayed if it's set to true and
if there is a beta message pass that in. It also tidies up some spacing
around the header.