Add Finder model
As we move away from Finder API we need to introduce a Finder object to
do the value to label conversion. This code is pretty much exactly the
same as what was in the API adapters other than it now always casts
value to an Array in line 28. This is because we no longer use single
selects in Finders.