Remove format specific Presenters
As we're moving to just having a Document model, we don't need all of
these Presenters. Apart from for DrugSafetyUpdate because it's a
special ❄️

DSU is unique in that they don't display any date metadata in the footer
if it's the first edition as we would then get a duplicate Published
(which is really `first_published_at`) and Published, which will change to
Updated when the document is updated. When it's not the first edition we
see Published (which is really `first_published_at`) and Updated, so
it's not a problem in that state.

For the default_date_metadata method, we return nothing if it's bulk
published or the first edition.