New DocumentPresenter
This new DocumentPresenter is now the Presenter for all document
formats.

The major change is now we pass the Finder to the class to get the
format name and the facets of the Finder, which now describe all the
Metadata that a document format may have. We then use the keys of these
facets to define methods which call `document.details.#{method}`.

We also use these facets to get a hash of the `extra_date_metadata`,
`filterable_metadata` and `extra_metadata`.