Configure asset host in development mode
When run under `govuk_setenv`, this configures the asset host to point
at the assets-origin on the dev VM.  This will give us better dev-prod
parity, and enable running this behind the router in dev.

See equiv on `frontend`:

alphagov/frontend@77bb2cb

I removed the manual setting of asset host `STATIC_DEV`, as this isn't
required and will work without - falling back. This is the way other apps
work, and @edds, who added it, is confident it's fine to remove