Upgrade Slimmer, don't explicitly set 'core_layout'
This has no effect on how the app runs, the same template will be used.

Slimmer 9.0.0 switched to core_layout as it's default layout option,
as it's the modern/recommend option. Apps explicitly setting a
layout should be outliers we plan on fixing.

Not setting the layout if we don't need to will make those outliers
easier to identify and simplifies the app controller code.