Remove summary padding, add metadata margin
We’re adding a default margin to the bottom of the `govuk-metadata`
component. If that were to go live as-is, then that margin would double
up with the summary’s padding.

Before we deploy the change to the metadata component we need to remove
the summary padding and add a temporary margin bottom so that
everything looks correct between deploys.