Add breadcrumbs to specialist content
Specialist frontend currently displays documents with a single link above the page title, linking to the parent finder. This commit replaces that link with a breadcrumb, using GOV.UK components. It links to the GOV.UK home page, and also the parent finder for the specified document, using the name of the finder as the link title.