Use the correct branch for schemas when running as a schema test
When a PR is raised against content schemas, we want to run the deployed-to-production branch of specialist-frontend against the branch of the PR in schemas - which comes in as the SCHEMA_BRANCH environment variable.

The contentSchemaDependency method will check out the `deployed-to-production` branch of schemas by default, unless the the SCHEMA_BRANCH is passed on as an argument.