Track standard content item analytics
Add govuk meta tags to each specialist document page. This enables
Google Analytics tracking of some content item properties like the
document type and linked organisations.

The main reason for this change is to track the document type (which is
tracked in the `govuk:format` meta tag against GA dimension 2). The
ranking of service standard reports (which are specialist documents) is
going to be changed in the search results, and this extra tracking will
help us analyse whether the changes to search have helped users find the
content they're looking for.